+++ read_input.h
@@ -15,7 +15,7 @@ typedef struct
 }molecule;
 
 void read_control(int* num_structures, int* Z, float* Zp_max,
-	float* volume_mean, float* volume_std, float *sr, int* max_attempts);
+	float* volume_mean, float* volume_std, float *sr, long* max_attempts);
 
 void read_geometry(molecule* mol);
 
@@ -24,6 +24,6 @@ void print_input_geometry(molecule* mol);
 void print_molecule(molecule *mol);
 
 void print_input_settings(int* num_structures, int* Z, float* Zp_max, 
-	float* volume_mean, float* volume_std, float *sr, int *max_attempts);
+	float* volume_mean, float* volume_std, float *sr, long *max_attempts);
 
 #endif
